Add files via upload
Think it works (no errors encountered during trade) but loses money on the second step. This is either: 1. due to small price fluctuations occurring between the initial and subsequent orders, or 2. due to the market trade just taking whatever offer is available which may not be the best value due to the arbitrage condition existing in the first place. Not really any way to speed up transactions besides better connection because the only calculations in the order loop require using variables generated in that loop. May try switching subsequent trades to limit in the next revision and see how it goes.